moved table repos into new folder for clarity
